add test Fixture without async, plus refactor to have test state class nested in test class to avoid clutter
