+++ ApprovalTests.Xunit2/CompatibilityWithXunit1FixutureTest.Xunit2ShouldWorkWithFixtureTest.approved.txt
@@ -1,0 +1,1 @@
+﻿{ 'result':'true' }
